Split out demo buttin and api menu
