+++ docs/site/src/components/ApiMenu.js
@@ -1,75 +1,53 @@
 // @flow weak
 
-import React, { PureComponent, PropTypes } from 'react';
-import IconButton from 'material-ui/IconButton';
-import { Menu, MenuItem } from 'material-ui/Menu';
-import MoreVertIcon from 'material-ui/svg-icons/more-vert';
-import { camelCase } from 'docs/site/src/utils/helpers';
+import React, { PropTypes } from 'react';
+import { camelCase, kebabCase } from 'docs/site/src/utils/helpers';
+import ApiIconMenu from './ApiIconMenu';
+import { apiMenus, componentMap } from './apiMenuData.js';
 
-export default class ApiMenu extends PureComponent {
-  static propTypes = {
-    className: PropTypes.string,
-    menuItems: PropTypes.array.isRequired,
-  };
+const ApiMenu = (props, context) => {
+  const path = window.location.hash.split('/');
 
-  static contextTypes = {
-    router: PropTypes.object.isRequired,
-  };
-
-  state = {
-    anchorEl: undefined,
-    open: false,
-  };
-
-  handleMenuClick = (event) => {
-    this.setState({
-      open: true,
-      anchorEl: event.currentTarget,
-    });
-  };
-
-  handleMenuItemClick = (event) => {
-    this.setState({ open: false });
-    this.context.router.push(`/component-api/${event.currentTarget.id}`);
-  };
-
-  handleMenuRequestClose = () => {
-    this.setState({
-      open: false,
-    });
-  };
-
-  render() {
-    const { menuItems, className } = this.props;
-
-    return (
-      <div>
-        <IconButton
-          contrast
-          onClick={this.handleMenuClick}
-          aria-owns="api-menu"
-          aria-haspopup="true"
-          className={className}
-        >
-          <MoreVertIcon />
-        </IconButton>
-        <Menu
-          id="api-menu"
-          anchorEl={this.state.anchorEl}
-          open={this.state.open}
-          onRequestClose={this.handleMenuRequestClose}
-        >
-          {menuItems.map((menuItem) => (
-            <MenuItem
-              key={menuItem}
-              id={menuItem}
-              onClick={this.handleMenuItemClick}
-            >
-              {camelCase(menuItem)}
-            </MenuItem>
-          ))}
-        </Menu>
-      </div>
-    );
+  // Return if we're on the home page.
+  if (path.length < 3) {
+    return null;
   }
-}
+
+  // component is the last part of the path
+  const component = path[path.length - 1];
+  let baseComponent;
+
+  // Check if the component is in the exceptions map
+  if (componentMap[component]) {
+    baseComponent = componentMap[component];
+  } else {
+    // Otherwise extract and pluralise the base component
+    baseComponent = `${(component.split('-'))[0]}s`;
+  }
+
+  let menuItems;
+
+  if (apiMenus[baseComponent]) {
+    menuItems = apiMenus[baseComponent];
+  } else {
+    const baseComponentName = camelCase(baseComponent.slice(0, -1));
+    menuItems = context.apiDocs
+      .filter((entry) => (entry.name.substr(0, baseComponentName.length) === baseComponentName))
+      .map((item) => (kebabCase(item.name)));
+  }
+
+  return (
+    menuItems &&
+    // Only show the menu on an api page if there's more than one entry
+    (path[1] !== 'component-api' || menuItems.length > 1) &&
+    <ApiIconMenu menuItems={menuItems} {...props} />
+  );
+};
+
+ApiMenu.contextTypes = {
+  apiDocs: PropTypes.array.isRequired,
+  demos: PropTypes.array.isRequired,
+  router: PropTypes.object.isRequired,
+};
+
+export default ApiMenu;
